requirements.txt
requests>=2.31.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
urllib3>=2.0.0
python-dotenv>=1.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
pandas-gbq>=0.19.0
db-dtypes>=1.1.1